+++ UnitUI/UIDamageText.cs
@@ -5,7 +5,7 @@ public sealed class UIDamageText : MonoBehaviour, IObjectPool
     public float LifeTime => 3f;
     [SerializeField] private TextMeshProUGUIEx _txtDamage;
 
-    public void Set(int damage)
+    public void Set(long damage)
     {
         _txtDamage.text = damage.ToStringMoney();
     }
